Complete basic single task multithreaded queue processing starting point
