+++ app/meticulous/_exceptions.py
@@ -1,0 +1,15 @@
+"""
+Meticulous errors
+"""
+
+
+class ProcessingFailed(Exception):
+    """
+    Raised if processing needs to go back to the main menu
+    """
+
+
+class NoRepoException(ProcessingFailed):
+    """
+    Raised if no repositories are available/selected
+    """
